Rename `run` module to `api`
