+++ pymagicc/api.py
